add naviagtion card
